Refactor remote installed plugin bundle sync
